Move some ORM things to a util module.
